DL-14089 All review comments addressed 2
