+++ project/AppDependencies.scala
@@ -5,7 +5,7 @@ object AppDependencies {
 
   val compile: Seq[ModuleID] = Seq(
     ws,
-    "uk.gov.hmrc"  %% "bootstrap-frontend-play-30"   % "9.0.0",
+    "uk.gov.hmrc"  %% "bootstrap-frontend-play-30"   % "9.2.0",
     "uk.gov.hmrc"  %% "domain-play-30"               % "10.0.0",
     "uk.gov.hmrc"  %% "play-partials-play-30"        % "10.0.0",
     "uk.gov.hmrc"  %% "play-frontend-hmrc-play-30"   % "10.5.0",
