+++ test/connectors/BackLinkCacheConnectorSpec.scala
@@ -35,7 +35,7 @@ class BackLinkCacheConnectorSpec extends PlaySpec with GuiceOneServerPerSuite  {
 
   val appConfig = app.injector.instanceOf[ApplicationConfig]
 
-  class Setup extends ConnectorTest1 {
+  class Setup extends ConnectorTest {
     val connector: BackLinkCacheConnector = new BackLinkCacheConnector(mockHttpClient, appConfig)
   }
